+++ stressinjector/__init__.py
@@ -1,0 +1,5 @@
+from stressinjector.cpu import CPUStress  # noqa: F401
+from stressinjector.memory import MemoryStress  # noqa: F401
+from stressinjector.url import URLStress  # noqa: F401
+
+version = "0.4a"